chore(phase-0): PostgreSQL + decimal money foundation
- БД переведена на PostgreSQL (миграции проходят, тесты зелёные)
- money/volume колонки float -> decimal(20,2): pv, bv,
  binary_small_branch_volume, rank_bonus_amount
- .gitignore расширен

+++ mh-calc-backend-main/Modules/Calculator/Database/Migrations/2024_09_24_095821_init_calculator.php
@@ -72,8 +72,8 @@ return new class extends Migration {
             $table->id();
             $table->foreignId('calculator_package_id')->constrained()->onDelete('cascade');
             $table->string('locale', 5)->index();
-            $table->float('pv', 20, 2)->default(0);
-            $table->float('bv', 20, 2)->default(0);
+            $table->decimal('pv', 20, 2)->default(0);
+            $table->decimal('bv', 20, 2)->default(0);
 
             $table->unique(['calculator_package_id', 'locale']);
         });
